CI fixes and improvements
* ensure incubating feature doesn't leak into normal builds
* disable incremental compilation on CI
